+++ module/Core/src/Model/DeviceType.php
@@ -9,18 +9,30 @@ enum DeviceType: string
 {
     case ANDROID = 'android';
     case IOS = 'ios';
+    case MOBILE = 'mobile';
+    case WINDOWS = 'windows';
+    case MACOS = 'macos';
+    case LINUX = 'linux';
+    case CHROMEOS = 'chromeos';
     case DESKTOP = 'desktop';
 
-    public static function matchFromUserAgent(string $userAgent): self|null
+    /**
+     * Determines which device types provided user agent matches. It could be more than one
+     * @return self[]
+     */
+    public static function matchFromUserAgent(string $userAgent): array
     {
         static $uaParser = new UserAgentParser();
         $ua = $uaParser->parse($userAgent);
 
         return match ($ua->platform()) {
-            Platforms::IPHONE, Platforms::IPAD => self::IOS, // Detects both iPhone and iPad (except iPadOS 13+)
-            Platforms::ANDROID => self::ANDROID, // Detects both android phones and android tablets
-            Platforms::LINUX, Platforms::WINDOWS, Platforms::MACINTOSH, Platforms::CHROME_OS => self::DESKTOP,
-            default => null,
+            Platforms::IPHONE, Platforms::IPAD => [self::IOS, self::MOBILE], // iPhone and iPad (except iPadOS 13+)
+            Platforms::ANDROID => [self::ANDROID, self::MOBILE], // android phones and android tablets
+            Platforms::LINUX => [self::LINUX, self::DESKTOP],
+            Platforms::WINDOWS => [self::WINDOWS, self::DESKTOP],
+            Platforms::MACINTOSH => [self::MACOS, self::DESKTOP],
+            Platforms::CHROME_OS => [self::CHROMEOS, self::DESKTOP],
+            default => [],
         };
     }
 }